Define variables in params class for strict mode
Without this change, the unix_sock_ro_perms and unix_sock_dir variables from
the params class are referenced but not defined.  This is a problem because
in strict mode this prevents the catalogue from compiling.

Note that this change is only targeted at the Debian OS family - it looks
like there are more variables undefined in other OS families, but I'm not
currently testing on those platforms thus have not attempted to fix them.